fix(core): 账本 #26——字符串字面量类型系统性丢失（PG unknown-literal 语义）
三处数值化源头同修：
① value_from_parser：引号串恒 Utf8（'1'/'1.5' 曾被数值化污染 TEXT 列；
  Number 字面量保持窄化定型 Int32/Int64）
② INSERT coerce_for_column 按目标列采纳：Utf8→数值/日期/时间戳/布尔
  按输入语法解析（22P02）；数值/布尔→TEXT 文本化；同族宽/窄化
  （Int32↔Int64 含 22003 溢出、整值 Float→Int）——此前 BIGINT 列存
  进 Int32 由 embed 启发式掩盖
③ embed to_result 按列元数据定型（此前对 text_rows 重猜 i64→f64→Utf8
  ——测试视角污染源，'1' 在 embed 里变回 Int64 的直接原因）

literal_typing 4 回归（TEXT 保型/按列采纳/22P02 错误/表达式拼接）；
期望更新 3 处（plan_cache Int32 窄化、ap_tp 下划线字面量、explain 行序）；
工作区 55 目标 0 panic + slt 26 全绿 + clippy 0

+++ crates/dendro-core/tests/plan_cache.rs
@@ -72,14 +72,17 @@ fn distinct_literal_text_distinct_results() {
         let r = conn.query(&format!("SELECT {n} AS n")).unwrap();
         assert_eq!(
             r.rows()[0][0],
-            dendro_core::types::SqlValue::Int64(n),
+            dendro_core::types::SqlValue::Int32(n as i32), // #26：Number 窄化定型（曾 embed 启发式 i64）
             "SELECT {n} 结果错乱"
         );
     }
     // 再跑一遍（命中路径）仍各归各位
     for n in 1..=5 {
         let r = conn.query(&format!("SELECT {n} AS n")).unwrap();
-        assert_eq!(r.rows()[0][0], dendro_core::types::SqlValue::Int64(n));
+        assert_eq!(
+            r.rows()[0][0],
+            dendro_core::types::SqlValue::Int32(n as i32)
+        );
     }
 }
 
@@ -91,7 +94,7 @@ fn cache_rejects_nothing_on_parse_error_miss() {
     assert!(conn.query("SELEC 1").is_err(), "重复报错行为一致");
     // 正确语句不受污染
     let r = conn.query("SELECT 1").unwrap();
-    assert_eq!(r.rows()[0][0], dendro_core::types::SqlValue::Int64(1));
+    assert_eq!(r.rows()[0][0], dendro_core::types::SqlValue::Int32(1)); // #26 同上
 }
 
 #[test]
